requirements.txt
requests==2.32.3
azure-storage-blob==12.23.1
python-dotenv==1.0.1